requirements.txt
asgiref==3.2.10
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cryptography==3.1.1
decouple==0.0.7
dj-database-url==0.5.0
Django==3.1.2
django-countries==6.1.3
django-rest-knox==4.1.0
djangorestframework==3.12.1
gunicorn==20.0.4
idna==2.10
pexpect==4.8.0
Pillow==7.2.0
ptyprocess==0.6.0
pycparser==2.20
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
Strip==0.0.1.dev14
stripe==2.55.0
urllib3==1.25.11
whitenoise==5.2.0
